include: Move line selection support
The include command entrypoint owns include workflows, while
selected-change structural staging already lives under commands.selection.

Line-selection staging helpers still live in the entrypoint module, so the
command flow sits behind transient batch staging, file-view checks,
source-line annotation, and baseline reference handling.

This commit begins separating include line handling by moving those helpers
into commands.selection.include_line_selection and routing include.py
through that module.

The next commit will move the include tests and architecture guard to the
new support-module boundary.
